fixing bug with transparent toolbar buttons in Partner Edit
see http://sourceforge.net/apps/mantisbt/openpetraorg/view.php?id=111

this is a problem of .Net.
see also http://bytes.com/topic/c-sharp/answers/551890-transparent-gif-using-bitmap-save [^]

in our code, in FormWriter.cs, we convert an .ico file to .gif. In that process, we lose the transparency.
Apparently, there are ways to workaround that, but they involve messing with the palette and copying the raw data, and it did not work on the first try.

Therefore the easiest solution is to manually convert the files with The Gimp from ico to gif files, and transparency works fine.