some adjustments to GUI
remove toolbar button for classic navigation. this can be enabled for users of legacy Petra 2.x, but will not be maintained in the public version
replace a few icons that have transparency, with the gif version of the icon
reorganize the navigation file, remove Demo section. we have now quite a number of working items already, no need to group them in Demo section