reports: income by fund
solve bug with printing lines without linefeed
uses parameters from the loaded settings file if they are not loaded from the GUI (yet). this allows column settings to be managed from the file.
clean up parameter gui for FD Income by fund report
only check a radio button if it is set in the settings file, do not uncheck a radio button. this allows for defaults to be set in the form (eg period range)
sort years available in ascending order
avoid too many warnings in DebugMode when parameter cannot be found. print only once for 20 occurances
fix timeout issue for postgresql
new version of Npgsql