fix issues with exporting and importing the full database with yml
see https://sourceforge.net/apps/mantisbt/openpetraorg/view.php?id=139
there was a problem with newline in attribute values, and also with # in values, which was treated as a comment.
also fixed some issue to allow login to the database even after a failed import. SSystemStatus is required for login.