improvements to bank import GUI, fix ignored transactions
display the sum of debits and credits
another tab for ignored transactions. this solves bug 183 with confusing behaviour after unmatching a transaction, it disappeared from unmatched transactions. now it is in the ignored list
add base and international currency code of the ledger to the ledger details which are cached