To verify that our Shepherd Framework inhertiance works I added two actual Shepherds, one for adding a Family and for adding a Church!
Mostly things worked out as we planned and were hoping for; just a few more Class Members and Methods needed to become 'protected' instead of 'private. 
In Method 'TPetraShepherdConcreteForm.Form_Load' we had a line that created an instance of 'TPetraShepherdFormLogic' and assigns this to FLogic; this was flawed as each instance of an actual Shepherd has its own Logic (which derives from 'TPetraShepherdFormLogic') and *that* must be instantiated and held in FLogic instead of an instance of the base class 'TPetraShepherdFormLogic'. This is so that the actual Shepherd has access to its own logic and to the base class logic in one object, FLogic, which in the actual Shepherd is casted to the Type of its own logic. In the actual Shepherd this I stored casted like that for convenience in a Field called 'FSpecificLogic'.

Added Logging to many Methods just to make it more transparent what gets called when and what Arguments are passed in to Constructors and Methods.