assigning null values or empty string to a column property should set the value to DBNull
this is for string, but also Nullable DateTime.
for string, if you read a column that has a null value in the database, you receive an empty string. It should work the same way when you assign an empty string, it should be set to null, to work symmetric.
For the nullable date, it should be possible to set a date that has no value.