nant recreateDatabase should not use hardcoded password and username for db user
when creating or recreating a database, the hardcoded username petraserver and password TOBESETBYINSTALLER have been used by the NAnt script.
This causes trouble when trying to start the petra server with nant startPetraServer, which writes the contents of the variables ${DBMS.UserName} and ${DBMS.Password} to the Server.config file.
Now recreateDatabase will also use the variables.