UserControls: always use a local instance of TValidationControlsDict rather than the instance of the Form (FPetraUtilsObject.ValidationControlsDict). Before that change it could happen under certain circumstancesthat  an Automatic Data Validation would be recorded twice.
Description of the specific circumstance: 
A Form could have added a Control for Data Validation for a DataColumn of a DataTable to FPetraUtilsObject.ValidationControlsDict. A UserControl that is hosted on that Form also has at least one Control that is bound to a *different* DataColum of the *same* DataTable, but is *does not have the same Control/DataColumn* than on the Form. In that circumstance the Automatic Data Validation for that DataTable is fired on the form level and on the usercontrol level because the DataTable is used on both. Now, if a data validation error occured on the Control on the form level, then the same error would be recorded again by the UserControl (although it doesn't have that Control!) because FPetraUtilsObject.ValidationControlsDict holds the Control for Data Validation for the DataColumn of the DataTable of the Form *already*, and the Automatic Data Validation for the same DataTable is fired in the UserControl *again* because at least one Control is bound to the same DataTable. Example: p_branch_name_c in Partner Edit's TopPart and tje Partner Details Tab UserControl on that Form: both validate PBank, and a data validation that fails on p_branch_name_c in Partner Edit's TopPart failed again in the Details Tab UserControl, resulting in the same Data Validation to be recorded twice!