Partner Find Screen - Bugfix for Clear button.
SingleDataRow["PartnerClass"] = FDefaultPartnerClass needs to be re-assigned again after a call to FPartnerClassDataTable.Rows.Clear(); - not really sure why this is necessary, but if this is not done then SingleDataRow["PartnerClass"] is ""..., which lead to the Bug. Wasn't needed in Petra 2.3.

Added DataSet FFindCriteriaDataSet, which aids in debugging the Find Screen by showing the Criteria DataTable's content that is going to be sent to the PetraServer if the CTRL key is pressed while the 'Search' Button is clicked. It was present in Petra 2.3's source code but we didn't have it in OpenPetra.