In the meantime the use of the CancelSaveException had unfortunately propagated to three forms instead of one!
In the Form 'Link Partner Cost Centre' CancelSaveException was just introduced by copy&paste and the handling of it was meaningless in that context, therefore the CancelSaveException was easy to remove. 
In the Form that I originally identified, 'GL Account Hierarchy', the process of getting rid of the CancelSaveException was more involved, but could be done without using the Data Validation Framework as MessageBoxes and cancelling of going ahead with updating the form proved to be more helpful in this case. In the third Form, 'GL Cost Centre Hierarchy', the process was very similar.

Finally, I removed the CancelSaveException from the windowTDS.cs Template and its Class got removed from PetraEditForm.cs so no-one can use it anymore (even by accident).