fix the problem with the NUnit test for banking details: now just using the date without time information for date_created and date_modified.
Problem in bankingdetails test:

creating a datarow, datecreated was set with DateTime.Now (with time).
submitted to database, database only stores the date.
then the datarow was marked for deletion.
a dataset was loaded with all banking details.
the deleted row was merged into that dataset.
because the dates differed, the result did not contain a deleted row, but a modified row.
this problem only occured on mono