WinForms Generator: Fix TabIndex numbers not being reset to 0 for each new file - runs into thousands across files.
I also took the opportunity to increase the TabIndex values not by 1 for each Control, but by 10. This helps in situations where the developer wants to have a custom Tab Order on a Form/UserControl and therefore would use the TabIndex Element in the YAML file to achieve that. Having 9 possibilities for reordering the TabIndex without necessarily needing to affect the TabIndex of other 'neighbouring' Controls will certainly be helpful! (Does anyone recall a similar situation with the BASIC programming language when one needed to insert code lines before existing code lines, but didnt' want to affect the line number of the existing code lines...?!)