Change to Method 'GetChangesTyped' of the auto-generated Typed DataSet Classes that are descendants of TTypedDataSet: check if the 'result' variable is not null before trying to attempt to call the Method InitVars() on it.
ChristopherT encountered situations where the 'result' variable is null in the Gift Batch screen  - for valid reasons - but got a got a System.NullReferenceException thrown in those cases when saving data in that screen. Chris says: 'Happend when I delete all gift details, add a new gift and immediately delete it'.

I checked that no code in OpenPetra relied on the 'side effect' of those auto-generated Methods throwing System.NullReferenceException in case the 'result' variable is null.