WinForms Generator: Allow creation of 'nice' horizontal pnlButtons and pnlDetailButtons Panels for Grids with improved Buttons on them, as shown in my demonstration in yesterdays' developers' video conference.
This change will affect only Forms and UserControls that have either a pnlButtons and pnlDetailButtons Control AND whose 'ControlsOrientation' is set to 'horizontal'. In the future all our pnlButtons and pnlDetailButtons will have that Element set to that value (and hence will have the changes applied that this commit implements), but at the moment only two Forms had that already and in those two Forms minimal layout changes to those Panels and their Buttons can be seen, but they are desirable:

  * \csharp\ICT\Petra\Client\MFinance\Gui\AP\APSupplierTransactions.yaml
  * \csharp\ICT\Petra\Client\MPartner\Gui\Extracts\ExtractFind.yaml

In all other Forms and UserControls there is no change to the layout because of this checkin, though the order of some Controls in the '*-Designer.cs' files is different in some files, but that has no consequence for the layout or functionality of those Forms.


This commit is also in preparation for the addition of the 'Filter' (and 'Find') functionality that is to be implemented in the future, which will then add the 'record counter' on the right and the ability to have a 'Filter' Button an the new horizontal pnlButtons and pnlDetailButtons Panels on those Forms/UserControls that will implement a Filter and/or Find functionality.

AlanP will follow up from this commit and change all YAML files of Forms/UserControls who have pnlButtons and pnlDetailButtons Panels on them (minor changes only) so that afterwards all Forms/UserControls will have horizontal pnlButtons and pnlDetailButtons Panels (with minor exceptions, such as Forms that only have 'up'/'down' Buttons for Grids - for those he will leave the pnlButtons and pnlDetailButtons Panels vertical).