Unit Tests: add new NAnt Task 'test-main-navigation-screens' that allows the running of the tests in the 'Tests.MainNavigationScreens' Namespace. That NAnt task is not included in the 'nant test' or 'nant test-all' commands as the new task runs for some time and we don't want to add that time to each usual test run.
Fixed two screens that refused to open when no Main Menu was present; not an unreasonable assumption if the regular OpenPetra Client runs, but when the NUnit Test runs there is no Main Form instance.

Also fixed TopDonorReport.ManualCode.cs which was assigning "0" to the .Text Propery of a TTxtNumericTextBox Control, which is now no longer possible due to the last commit of Timotheus Pokorra. Now assigning 0 to the NumberValueInt Property of that Control.