The new SourceGrid.dll that I checked in in my previous branch caused some Finance screens to fail to load correctly - notably GLBatch and Recurring GL Batch.  The basic cause was down to the rowChange event now being based on the Selection.Focus method which does not normally give rise to events before the grid has been shown.  So in the previous commit I changed the SelectRowInGrid call to be in RunOnceOnActivation.  This is not really any good because a lot of the Finance code wants to set up user controls etc in their constructors and rely on the fact that we know the details for the row.  So the simplest fix was to change the grid Focus method so that we still fire the SelectionChanged event even before the grid is active.  I will remove the SelectRowInGrid call from the RunOnceOnActivation method in the templates and insert it at the end of the constructor.
In addition the Finance modules subscribe to some other grid events that I had not checked and which can produce a stack overflow.  So I had to add a specific suppress events parameter to one more call in the grid internal code.

Finally I modified my row selection code to preserve the active column when selecting a row.  This became necessary for the Analysis Values grid on the GL Transactions screen - which relies on being able to focus clolumn 1 for the edit-in-place combo box.