Some changes to GL Batch screens to work with the new SourceGrid.  Some of these screens called grid.Selection.Focus directly and that is not now possible without causing a cascade of events and a stack overflow.  Instead it is necessary to call ReselectGridRowAfterSort in order to avoid this.
Also added a status bar text item for the analysisAttributes grid so it is easier to see that the grid has the focus when tabbing around controls.

Also added some code to prevent un-necessary events firing off during LoadBatches.  This has greatly reduced the time taken to load the GL Batch screen.