Ledger Creation: major parts of the Ledger Creation program code (Method 'CreateNewLedger' in GL.Setup.cs) were run WITHOUT A DB TRANSACTION! That is very dangerous and must therefore not be done. Rectified!
Reason why this caused the Bug: If no Partner exists for the Ledger, a Partner is created. When the Partner is created, a check on whether an 'empty Location record' (the one with "No valid address on file" in the Street Name) exists yet FOR THE COUNTRY that got selected in the Create Ledger screen. When such a Location doesn't exist yet, the program code creates that record. To do so it needs to get a DB Sequence value from the DB using the GetNextSequenceValue Method so it can assign a LocationKey, and that works on PostgreSQL, but breaks on SQLite. The difference lies in the implementation of those Methods - for SQLite it needs an active DB Transaction, but not for PostgreSQL.