UpdateExtractChangeSubscriptionDialog: * Fixes regarding Date Checks. * Further fixes for more checks that were either not quite done like in the corresponding 4GL screen, or were missing altogether.
UC_PartnerSubscriptions.cs: None of the logic contained in that class was used anymore; it was a remnant of Petra 2.3's Subscriptions UserControl which worked rather different than the UserControl in OpenPetra. I emptied the Class, but placed a new shared Method for checking of the validity of the selected Publication in a ComboBox. That check is needed both in the Partner Edit Screen's Subscriptions Tab and in the UpdateExtractChangeSubscriptionDialog.

Template window.cs: Added the missing option to process Data Validation errors without displaying them.