Wide-reaching fix for the problem that the AVerificationResult Argument of the auto-generated SubmitChanges Methods of a Typed DataSet were not reliable: * Changed the implementation of the SubmitChanges Methods in TypedDataAccess.cs so that they no longer return AVerificationResult Arguments *at all*. That was achieved by no longer wrapping Exceptions into TVerificationResult objects, instead we now simply let the Exceptions happen. That is cleaner and in that way it is much less likely that such Exceptions will go unnoticed, which was possible before if the caller (or the caller chain) accidentally discarded the AVerificationResult that was returned by those Methods! * That allowed for the change that the auto-generated SubmitChanges Methods of a Typed DataSet no longer have a AVerificationResult Argument *at all* (those changes are in DataSetAccess.cs), and this fixes Bug #721.
Those changes necessitated wide-reaching implementation changes across the OpenPetra codebase, which are part of this commit. As much as possible those have been thought through carefully, though the committer can not know about all the details of all the call chains of the code that was affected by those changes...

All the code compiles 
The NUnit Tests and also the Main Menu Nunit Tests pass.