A number of small fixes and changes to the code to make these screens work correctly after the integration of filter/find.  There were a number of issues surrounding deleting records and renumbering transactions.  Some lines of code needed re-ordering so as to work correctly.  In a few cases there were lines that were missing.  I have now tested the deletion so that I am sure that renumbering works and the storage of the last transaction number also works right.
The button for the DeleteAllGifts method is now disabled if the discretionary part of the filter has been set, since the method really does delete all the gifts, not just the ones that are being shown by the filter.  In the future we may decide that we will delete the complete filtered subset - but that would involve alll the renumbering routines to be applied.