Added a property to the paged data grid - MinimumPageSize.  This defaults to 1 (row) but, when set to a number such as 200 specifies that the grid should load 200 rows (if available) rather than the number of visible rows (say typically 25).  beyond that the grid works exactly as before.  If the viewport needs to show a row that has not been retrieved it will get the next set of rows (200 more in this example).
This is how the Accounts Payable screen works now.  It specifies 200 rows and then when they have been retrieved it requests to get all the remaining rows, so the grid is fully populated at the outset although it displays the initial rows very quickly because 200 rows is not very many over a network connection.