This turned out to be all that was necessary to get grid auto-sizing (and stretch) to work throughout the OP application.  Essentially I added just one line to the grid wrapper OnResize event handler and this did everything that we need!  Columns now automatically generate their width from the first 200 (by default) rows, optionally including the header text in the calculation.  Furthermore, the great news is that this is a better way of working than the (temporary) fix that I did a few months ago for certain selected screens.  That was triggered by trying to minimise the time taken to display a grid with thousands of rows.  Also at that time there seemed to be a strange behaviour where repeated maximise and restore of a window went progressively slower!
This new checkin seems to work ok with all these issues and just works everywhere for all screens.

When the filter panel is toggled the grid now resizes (which is great).  When the filter panel is shown the header text is not included in the column width calculation.  When the filter is not shown, it is included.