Added new code to show scrollbars in Filter/Find panel when used on smaller screens - or just when the user wants the OP screen to be small for some reason.
For some reason that I can't explain the Filter panel just gets avertical scrollbar (which I like) but the Find panel gets both vert and horiz!  I ended up deciding that it was probably the Find panel that Windows did what it meant to do and for some reason the Filter panel was the 'exception'.  Anyway - I am committing as-is.  Two scrollbars aren't too bad.