Fixed the long outstanding issue on Local Partner Data where the grid displays text like float and boolean whereas the options we want to see (and which match what the details panel has been written to show) are Number and Yes/No.
I had thougtht that this would be done with a TypeConverter, so I started with that - but the grid would not use it because there is no need to convert text to text.  Then it looked like it could be done with an Editor (or Model or View or Controller) which might have been possible but in those classes, while it is possible to read what is in the cell iut is not possible to write it.

So in the end I realised it would be simple to do what has been done already in this screen and that is to add a calculated column to the data set and point the grid column to that.  Simples!!