3122: https://tracker.openpetra.org/view.php?id=4896 The CH office encountered a problem with importing a file containing updates to Event Codes. It included 4 partners with a partner key ending in 0, including one which was 1700000. OP optionally treats keys ending in 0 as referring to a 'range' of that 'decade' - so 1700000 becomes 'between 1700000 and 1799999'. This played havoc with the logic since the number of codes returned is limited to 50 and we would have got hundreds and anyway we did not get back the partner key we were looking for.
So I have added a bool method parameter to specify that we want an exact match, which will return either 1 row or possibly none. The method in question is actually only used by this one import method, so I could have just always set it to Exact Match, but at least this way the option is clear. Normal Partner/Find does not use this call but does make use of the user option to match exact - which changes the behaviour of the Partner/Find screen. OP does not provide a GUI for setting the user preference though so it will be rarely used.