3289: Changes to the Main Navigation Screens Tests because the way that we handle security exceptions has changed - we now have a real Exception that needs handling.
Also two of the screens allowed DEMO to open them without SYSMAN permissions.  That is now fixed and we test for it too.

There are two more screens that allow DEMO to open them (Country and Language Specifics) but they involve cacheable tables and we can't do anything about that yet.

Finally the Open all screens test now correctly discovers that Country requires PTNRADMIN access and therefore is not opened.  A new category has been created for this.

Tags: BETA1_0.6.16